refactor: 공통 컴포넌트 분리 refactor: useState 사용 하지 않고 라우팅으로 변경
