+++ src/App.js
@@ -6,11 +6,12 @@ import { ThemeProvider } from 'styled-components';
 import FloatingButton from './components/FloatingButton';
 import Footer from './components/Footer/Footer';
 import Header from './components/Header';
-import ApplySuccessPage from './pages/ApplyCheckResult';
+import ApplyFailPage from './pages/ApplyFail';
 import ApplyFinishPage from './pages/ApplyFinish';
 import ApplyInfoPage from './pages/ApplyInfo';
 import ApplyMainPage from './pages/ApplyMain';
 import ApplyNotFoundPage from './pages/ApplyNotFound';
+import ApplySuccessPage from './pages/ApplySuccess';
 import ApplyWritePage from './pages/ApplyWrite';
 import GlobalStyle from './styles/GlobalStyle';
 import { Theme } from './styles/Theme';
@@ -33,7 +34,8 @@ function App() {
           <Route path="/write" element={<ApplyWritePage />} />
           <Route path="/finish" element={<ApplyFinishPage />} />
           <Route path="/*" element={<ApplyNotFoundPage />} />
-          <Route path="/result" element={<ApplySuccessPage />} />
+          <Route path="/success" element={<ApplySuccessPage />} />
+          <Route path="/fail" element={<ApplyFailPage />} />
         </Routes>
         <FloatingButton />
         <Footer />
