+++ src/pages/ApplyFail/index.js
@@ -1,21 +1,33 @@
-import React from 'react';
-
 import styled from 'styled-components';
 
-import ApplyTitleBox from './ApplyTitleBox';
+import ApplyTitleBox from '../../components/ApplyResult/ApplyCheckTitleBox';
 
-const FailContent = () => {
+const ApplyFailPage = () => {
   return (
-    <>
-      <ApplyTitleBox success={false} />
+    <ApplyCheckTopBox>
+      <ApplyTitleBox result={'FAIL'} />
       <ApplyFailResultTop>
         김땡땡님의 지원이&nbsp;
         <RedBottonLine>확인되지 않았습니다</RedBottonLine>
       </ApplyFailResultTop>
       <ApplyFailResultBotton>페이지 오른쪽 하단의 채널톡을 통해 문의해주세요</ApplyFailResultBotton>
-    </>
+    </ApplyCheckTopBox>
   );
 };
+
+const ApplyCheckTopBox = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  height: calc(100vh - 56px - 160px);
+  @media ${({ theme }) => theme.devices.TABLET} {
+    height: calc(100vh - 70px - 170px);
+  }
+  @media ${({ theme }) => theme.devices.DESKTOP} {
+    height: calc(100vh - 70px - 232px);
+  }
+`;
 
 const ApplyFailResultTop = styled.div`
   font-size: 14px;
@@ -62,4 +74,4 @@ const RedBottonLine = styled.span`
   }
 `;
 
-export default FailContent;
+export default ApplyFailPage;
